uses benchmark to show times for only the queries
Now, I want to do the database calls as a rake task. However, the current views benchmark the queries themselves and without using the output in a useful way i.e. by printing them in the view.

I have 3 case scenarios, because the speed of the task depends on the use case. I decided to benchmark calls from one model, two models, and three models because that seems like the most common scenarios.